# aspnetcore/src/IO.Swagger/Controllers/TagsValidationApi.cs
/*
 * Digitalization of Asset Counting Web Api
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TagsValidationApiController : ControllerBase
    { 
        /// <summary>
        /// Returns whether specified tag exists and is already in use or not.
        /// </summary>
        
        /// <param name="tag">define the tag that shall be validated</param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/tags/validation")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("Get")]
        [SwaggerResponse(statusCode: 200, type: typeof(TagValidation), description: "Success")]
        public virtual IActionResult Get([FromQuery]string tag)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(TagValidation));

            string exampleJson = null;
            exampleJson = "{\n  \"isValid\" : true,\n  \"inUse\" : true,\n  \"tag\" : \"tag\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<TagValidation>(exampleJson)
            : default(TagValidation);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}