# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v1",
    "title" : "Digitalization of Asset Counting Web Api"
  },
  "security" : [ {
    "Bearer" : [ ],
    "Basic" : [ ]
  } ],
  "paths" : {
...
    "/api/sapinventoryvariants" : {
      "get" : {
        "tags" : [ "SapInventoryVariants" ],
        "summary" : "Returns list of Inventory Variants. This query can be paged. If page and pageSize are omitted,\r\nthe full list is going to be returned. When pageSize and page specified,\r\npageCount and currentPage will be returned on first call.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "define the page that this query will return",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "define the number of items each page will contain if records are available",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SapInventoryVariant"
              },
              "uniqueItems" : false
            }
          }
        }
      },
      "post" : {
        "tags" : [ "SapInventoryVariants" ],
        "summary" : "Gets the SapInventoryVariant specified by inventory variant name.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "name" : "sapInventoryVariantName",
          "in" : "query",
          "description" : "the name of the SapInventoryVariant to be retrieved",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SapInventoryVariant"
              },
              "uniqueItems" : false
            }
          }
        }
      }
    },
...
    "/api/tags/validation" : {
      "get" : {
        "tags" : [ "TagsValidation" ],
        "summary" : "Returns whether specified tag exists and is already in use or not.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "name" : "tag",
          "in" : "query",
          "description" : "define the tag that shall be validated",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/TagValidation"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Bearer" : {
      "description" : "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "TagValidation" : {
      "type" : "object",
      "properties" : {
        "tag" : {
          "type" : "string"
        },
        "inUse" : {
          "type" : "boolean"
        },
        "isValid" : {
          "type" : "boolean"
        }
      },
      "example" : {
        "isValid" : true,
        "inUse" : true,
        "tag" : "tag"
      }
    },
...
    "SapInventoryVariant" : {
      "type" : "object",
      "properties" : {
        "inventoryVariantName" : {
          "type" : "string",
          "maxLength" : 255
        },
        "columnName" : {
          "type" : "string",
          "maxLength" : 50
        },
        "value" : {
          "type" : "string",
          "maxLength" : 4000
        },
        "valueRange" : {
          "type" : "string",
          "maxLength" : 4000
        },
        "operator" : {
          "type" : "string",
          "maxLength" : 2
        },
        "includeExclude" : {
          "type" : "string",
          "maxLength" : 1
        },
        "sourceSystem" : {
          "type" : "string",
          "maxLength" : 3
        }
      },
      "example" : {
        "valueRange" : "valueRange",
        "sourceSystem" : "sourceSystem",
        "inventoryVariantName" : "inventoryVariantName",
        "value" : "value",
        "includeExclude" : "includeExclude",
        "operator" : "operator",
        "columnName" : "columnName"
      }
    }
  }
}